host/tinkoff/requirements.txt
Package             Version
------------------- --------
cachetools          5.3.1
deprecation         2.1.0
grpcio              1.57.0
packaging           23.1
pip                 23.2.1
protobuf            4.24.0
PyMySQL             1.1.0
python-dateutil     2.8.2
redis               5.0.0
setuptools          65.5.0
six                 1.16.0
tinkoff             0.1.1
tinkoff-investments 0.2.0b59
